requirements.txt
# RealtyGenie Pro - AI Property Marketing System
# Core Dependencies

# Deep Learning & AI
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
safetensors>=0.3.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Web Framework
streamlit>=1.28.0

# Visualization
plotly>=5.15.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2

# Optional: For better performance
accelerate>=0.20.0